+++ lib/smmorphwavsource.cc
@@ -3,6 +3,8 @@
 #include "smmorphwavsource.hh"
 #include "smmorphplan.hh"
 #include "smleakdebugger.hh"
+#include "smproject.hh"
+#include "smzip.hh"
 
 using namespace SpectMorph;
 
@@ -26,7 +28,9 @@ MorphWavSource::MorphWavSource (MorphPlan *morph_plan) :
   auto position = add_property (&m_config.position_mod, P_POSITION, "Position", "%.1f %%", 50, 0, 100);
   position->modulation_list()->set_compat_type_and_op ("position_control_type", "position_op");
 
-  connect (morph_plan->signal_operator_removed, this, &MorphWavSource::on_operator_removed);
+  UserInstrumentIndex *user_instrument_index = morph_plan->project()->user_instrument_index();
+  connect (user_instrument_index->signal_instrument_updated, this, &MorphWavSource::on_instrument_updated);
+  connect (user_instrument_index->signal_bank_removed, this, &MorphWavSource::on_bank_removed);
 }
 
 MorphWavSource::~MorphWavSource()
@@ -50,17 +54,53 @@ MorphWavSource::object_id()
 }
 
 void
-MorphWavSource::set_instrument (int instrument)
+MorphWavSource::set_bank_and_instrument (const string& bank, int instrument)
 {
-  m_instrument = instrument;
+  instrument = std::clamp (instrument, 1, 128);
+  if (m_bank != bank || m_instrument != instrument)
+    {
+      m_bank = bank;
+      m_instrument = instrument;
 
-  m_morph_plan->emit_plan_changed();
+      Project *project = morph_plan()->project();
+      Instrument *instrument = project->get_instrument (this);
+      UserInstrumentIndex *user_instrument_index = project->user_instrument_index();
+
+      Error error = instrument->load (user_instrument_index->filename (m_bank, m_instrument));
+      if (error)
+        {
+          /* most likely cause of error: this user instrument doesn't exist yet */
+          instrument->clear();
+        }
+      project->rebuild (this);
+
+      signal_labels_changed();
+
+      m_morph_plan->emit_plan_changed();
+    }
+}
+
+void
+MorphWavSource::on_bank_removed (const string& bank)
+{
+  if (bank == m_bank)
+    {
+      m_bank = ""; // force re-read
+
+      set_bank_and_instrument (MorphWavSource::USER_BANK, 1);
+    }
 }
 
 int
 MorphWavSource::instrument()
 {
   return m_instrument;
+}
+
+string
+MorphWavSource::bank()
+{
+  return m_bank;
 }
 
 void
@@ -75,6 +115,32 @@ string
 MorphWavSource::lv2_filename()
 {
   return m_lv2_filename;
+}
+
+void
+MorphWavSource::on_instrument_updated (const std::string& bank, int number, const Instrument *new_instrument)
+{
+  if (bank == m_bank && number == m_instrument)
+    {
+      auto project  = m_morph_plan->project();
+      Instrument *instrument = project->get_instrument (this);
+
+      if (new_instrument->size())
+        {
+          ZipWriter new_inst_writer;
+          new_instrument->save (new_inst_writer);
+          ZipReader new_inst_reader (new_inst_writer.data());
+          instrument->load (new_inst_reader);
+        }
+      else
+        {
+          instrument->clear();
+        }
+      project->rebuild (this);
+      project->state_changed();
+
+      signal_labels_changed();
+    }
 }
 
 const char *
@@ -97,6 +163,7 @@ MorphWavSource::save (OutFile& out_file)
   out_file.write_int ("object_id", m_config.object_id);
   out_file.write_int ("instrument", m_instrument);
   out_file.write_string ("lv2_filename", m_lv2_filename);
+  out_file.write_string ("bank", m_bank);
 
   return true;
 }
@@ -132,6 +199,10 @@ MorphWavSource::load (InFile& ifile)
             {
               m_lv2_filename = ifile.event_data();
             }
+          else if (ifile.event_name() == "bank")
+            {
+              m_bank = ifile.event_data();
+            }
           else
             {
               g_printerr ("bad string\n");
@@ -146,11 +217,6 @@ MorphWavSource::load (InFile& ifile)
       ifile.next_event();
     }
   return true;
-}
-
-void
-MorphWavSource::on_operator_removed (MorphOperator *op)
-{
 }
 
 MorphOperator::OutputType