+++ NEWS.md
@@ -3,6 +3,7 @@
 #### New features
 * Support click & drag in instrument editor sample to scroll & zoom (#22).
 * Support stereo to mono conversion for files added in instrument editor (#14).
+* Support multiple banks for wav sources / instrument editor.
 
 #### Fixes
 * Fix crash if instrument editor is closed without any samples